Move ActiveVotingProgress to components
