+++ src/pages/Vote/GovernanceProposal/styles.ts
@@ -62,11 +62,6 @@ export const useStyles = () => {
     `,
 
     /* StatusCard styles */
-    votesWrapper: css`
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-    `,
     voteRow: css`
       display: flex;
       justify-content: space-between;
